minor code cleanup

+++ Fluid-Particle-Simulator/src/physics/SolveCollision.h
@@ -4,10 +4,10 @@
 // Solve collision between particle (particleA) and simulation 
 void SolveCollisionBorder(Particle& particleA,
     const std::vector<glm::vec2> simBounds,
-    unsigned int particleRadius);
+    float particleRadius);
 
 // Solve collision between particle A and particle B.
 // At the moment this uses a naive approach but will be improved
 void SolveCollisionParticle(Particle& particleA, Particle& particleB,
     const std::vector<glm::vec2> simBounds,
-    unsigned int particleRadius);
+    float particleRadius);
